Added a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-int i = 1, j = 2, k, m;
+double i = 1, j = 2, k, m;
 printf("%d, ", i);
 printf("%d, ", j);
 for (k = 0; k < 48; k++)
